requirements.txt
langgraph==0.4.5
langchain==0.3.25
langchain-openai==0.3.17
langgraph-checkpoint-postgres==2.0.21
fastapi==0.115.12
uvicorn==0.32.1
redis==6.2.0
concurrent-log-handler==0.9.28
pydantic-settings==2.5.2
python-dotenv==1.0.1
langchain-mcp-adapters==0.1.1